leader-toolbox/memory-py/requirements.txt
faiss-cpu==1.13.2
fastapi>=0.128.4
numpy<2
pytest>=9.0.2
sentence-transformers>=5.2.2
uvicorn[standard]>=0.40.0

# Memory system dependencies
watchdog>=4.0.0          # File watching
openai>=1.0.0            # OpenAI embeddings
google-generativeai>=0.8.0  # Gemini embeddings
voyageai>=0.2.0          # Voyage embeddings
click>=8.0.0             # CLI interface
pydantic>=2.0.0          # Data validation (already included via FastAPI)
sqlite-vec==0.1.6        # Vector extension (optional)